test(e2e): fix cypress test for nx
