--- apps/insight-viewer-docs-e2e/src/e2e/app.cy.ts
@@ -1,13 +1,0 @@
-import { getGreeting } from '../support/app.po';
-
-describe('insight-viewer-docs', () => {
-  beforeEach(() => cy.visit('/'));
-
-  it('should display welcome message', () => {
-    // Custom command example, see `../support/commands.ts` file
-    cy.login('my-email@something.com', 'myPassword');
-
-    // Function helper example, see `../support/app.po.ts` file
-    getGreeting().contains('Welcome insight-viewer-docs');
-  });
-});
