+++ apps/insight-viewer-docs-e2e/src/e2e/interaction.custom.cy.ts
