+++ apps/insight-viewer-docs-e2e/src/e2e/interaction.cy.ts
