+++ apps/insight-viewer-docs-e2e/src/e2e/measurement.viewer.cy.ts
@@ -1,7 +1,7 @@
 import { setup } from '../support/utils'
 import { VIEWPORT_WIDTH, VIEWPORT_HEIGHT, $LOADED } from '../support/const'
-import { RULER_MEASUREMENTS } from '../../mocks/ruler'
-import { CIRCLE_MEASUREMENTS } from '../../mocks/circles'
+import { RULER_MEASUREMENTS } from '../../../insight-viewer-docs/mocks/ruler'
+import { CIRCLE_MEASUREMENTS } from '../../../insight-viewer-docs/mocks/circles'
 
 describe(
   'Measurement Viewer',
