+++ apps/insight-viewer-docs-e2e/src/support/mockContours.ts
