+++ package.json
@@ -45,9 +45,14 @@
     "@nrwl/react": "14.5.6",
     "@nrwl/web": "14.5.6",
     "@nrwl/workspace": "14.5.6",
+    "@percy/cli": "^1.9.1",
+    "@percy/cypress": "^3.1.2",
     "@testing-library/react": "^12.1.5",
+    "@types/cornerstone-core": "^2.3.0",
     "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
+    "@types/point-in-polygon": "^1.1.1",
+    "@types/polylabel": "^1.0.5",
     "@types/react": "17.0.0",
     "@types/react-dom": "17.0.0",
     "@typescript-eslint/eslint-plugin": "^5.29.0",
@@ -65,14 +70,11 @@
     "jest": "27.5.1",
     "nx": "14.5.6",
     "prettier": "^2.6.2",
+    "react-is": "^17.0.2",
     "react-test-renderer": "17.0.2",
     "ts-jest": "27.1.4",
     "ts-node": "10.9.1",
-    "typescript": "~4.7.2",
-    "@types/cornerstone-core": "^2.3.0",
-    "@types/point-in-polygon": "^1.1.1",
-    "@types/polylabel": "^1.0.5",
-    "react-is": "^17.0.2"
+    "typescript": "~4.7.2"
   },
   "msw": {
     "workerDirectory": "apps/insight-viewer-docs/public"
